+++ CHANGELOG.md
@@ -25,6 +25,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Fixed
+- Fixed a bug in the `bos-operations` stage of `sat bootsys` where a Bad Request
+  error could appear in the output when checking node states.
+- Fixed a bug where the `bos-operations` stage of `sat bootsys` would not correctly
+  wait for a shutdown or boot operation to complete with BOS v1 and certain versions
+  of the `kubectl` command.
+
 ## [3.21.2] - 2023-02-13
 
 ### Fixed